Susanqueen (#6)
* King Tide reference

Updated King Tide reference to 2.5 instead of 2.0.

* const to var

Changed several items from type const to type var. Basically, everything that is a var now was initially a const.

* Labels

Added Text to Station ID at bottom of page

* Made Space for Label

Changed the dimensions of the map to accommodate the legend at the bottom.

* Added Legend Styles

Added legend styles.

* Added Legend to HTML

Added #leyend to .html

* Added med2 legend

Added a style for a med2 legend.

* Edited Legend

Changed legend from 3 to 4 states to accommodate once-a-day high tides and twice-a-day high tides.